+++ library/android-mvc-test/src/main/java/com/shipdream/lib/android/mvc/view/viewpager/ViewPagerTestActivity.java
@@ -16,6 +16,8 @@
 
 package com.shipdream.lib.android.mvc.view.viewpager;
 
+import android.content.Intent;
+
 import com.shipdream.lib.android.mvc.view.MvcActivity;
 import com.shipdream.lib.android.mvc.view.MvcFragment;
 
@@ -41,4 +43,8 @@ public class ViewPagerTestActivity extends MvcActivity {
         }
     }
 
+    void launchAnotherActivity() {
+        Intent intent = new Intent(this, ViewPagerTestActivityTop.class);
+        startActivity(intent);
+    }
 }