+++ library/android-mvc-test/src/main/java/com/shipdream/lib/android/mvc/view/viewpager/ViewPagerTestActivityTop.java
@@ -1,0 +1,14 @@
+package com.shipdream.lib.android.mvc.view.viewpager;
+
+import android.app.Activity;
+import android.os.Bundle;
+
+import com.shipdream.lib.android.mvc.view.test.R;
+
+public class ViewPagerTestActivityTop extends Activity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_view_pager_top);
+    }
+}